Refactor Button files to separate common files
Separate StyledButton and ButtonProps into different files so that the button variations do not have to import these items from BasicButton
