+++ src/Button/ButtonProps.tsx
@@ -1,0 +1,12 @@
+import { ReactNode, MouseEventHandler } from 'react';
+
+export interface ButtonProps {
+  /** Text or components to be displayed on the button */
+  children?: ReactNode;
+  /** Function to call on click event */
+  clickHandler: MouseEventHandler;
+  /** If true, button won't fire */
+  disabled?: boolean;
+  /** Allows you to pass in style properties for the element */
+  theme?: object;
+}
